[merge] fix merge
